// food-channel-processor/src/v1/grpc/interfaces/branch-channel-food-branch-map.ts
// Code generated by protoc-gen-ts_proto. DO NOT EDIT.
// versions:
//   protoc-gen-ts_proto  v1.176.0
//   protoc               v3.5.1
// source: branch-channel-food-branch-map.proto

/* eslint-disable */
import { Metadata } from "@grpc/grpc-js";
import { GrpcMethod, GrpcStreamMethod } from "@nestjs/microservices";
import { Observable } from "rxjs";

export const protobufPackage = "vn.techres.microservice.grpc.java.dashboard.branch_channel_food_branch_map";

export interface UpdateBranchMapWhenUnconnectionRequest {
  restaurant_id: number;
  restaurant_brand_id: number;
  channel_order_food_id: number;
}

export interface UpdateBranchMapWhenUnconnectionTokenRequest {
  restaurant_id: number;
  restaurant_brand_id: number;
  channel_order_food_id: number;
  channel_order_food_token_id: number;
}

export interface BaseResponse {
  status: number;
  message: string;
  data: Empty | undefined;
}

export interface Empty {
}

/** ----------- updateSettingBranch ---------- */
export interface UpdateSettingBranchRequest {
  branch_id: number;
  list_remove: string;
}

/** ----------- updateSettingFoodChannelRestaurantBrand ---------- */
export interface UpdateSettingFoodChannelRestaurantBrandRequest {
  channel_order_food_token_ids: string;
}

export const VN_TECHRES_MICROSERVICE_GRPC_JAVA_DASHBOARD_BRANCH_CHANNEL_FOOD_BRANCH_MAP_PACKAGE_NAME =
  "vn.techres.microservice.grpc.java.dashboard.branch_channel_food_branch_map";

export interface BranchChannelFoodBranchMapServiceClient {
  updateBranchMapWhenUnconnection(
    request: UpdateBranchMapWhenUnconnectionRequest,
    metadata?: Metadata,
  ): Observable<BaseResponse>;

  updateBranchMapWhenUnconnectionToken(
    request: UpdateBranchMapWhenUnconnectionTokenRequest,
    metadata?: Metadata,
  ): Observable<BaseResponse>;

  updateSettingBranch(request: UpdateSettingBranchRequest, metadata?: Metadata): Observable<BaseResponse>;

  updateSettingFoodChannelRestaurantBrand(
    request: UpdateSettingFoodChannelRestaurantBrandRequest,
    metadata?: Metadata,
  ): Observable<BaseResponse>;
}

export interface BranchChannelFoodBranchMapServiceController {
  updateBranchMapWhenUnconnection(
    request: UpdateBranchMapWhenUnconnectionRequest,
    metadata?: Metadata,
  ): Promise<BaseResponse> | Observable<BaseResponse> | BaseResponse;

  updateBranchMapWhenUnconnectionToken(
    request: UpdateBranchMapWhenUnconnectionTokenRequest,
    metadata?: Metadata,
  ): Promise<BaseResponse> | Observable<BaseResponse> | BaseResponse;

  updateSettingBranch(
    request: UpdateSettingBranchRequest,
    metadata?: Metadata,
  ): Promise<BaseResponse> | Observable<BaseResponse> | BaseResponse;

  updateSettingFoodChannelRestaurantBrand(
    request: UpdateSettingFoodChannelRestaurantBrandRequest,
    metadata?: Metadata,
  ): Promise<BaseResponse> | Observable<BaseResponse> | BaseResponse;
}

export function BranchChannelFoodBranchMapServiceControllerMethods() {
  return function (constructor: Function) {
    const grpcMethods: string[] = [
      "updateBranchMapWhenUnconnection",
      "updateBranchMapWhenUnconnectionToken",
      "updateSettingBranch",
      "updateSettingFoodChannelRestaurantBrand",
    ];
    for (const method of grpcMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcMethod("BranchChannelFoodBranchMapService", method)(constructor.prototype[method], method, descriptor);
    }
    const grpcStreamMethods: string[] = [];
    for (const method of grpcStreamMethods) {
      const descriptor: any = Reflect.getOwnPropertyDescriptor(constructor.prototype, method);
      GrpcStreamMethod("BranchChannelFoodBranchMapService", method)(constructor.prototype[method], method, descriptor);
    }
  };
}

export const BRANCH_CHANNEL_FOOD_BRANCH_MAP_SERVICE_NAME = "BranchChannelFoodBranchMapService";